added behaviour description into README and change examples for float as return value

+++ examples/BH1750onetime/BH1750onetime.ino
@@ -33,7 +33,7 @@ void setup(){
 
 void loop() {
 
-  uint16_t lux = lightMeter.readLightLevel();
+  float lux = lightMeter.readLightLevel();
   Serial.print("Light: ");
   Serial.print(lux);
   Serial.println(" lx");
